demo-server/requirements.txt
django==3.0.1
django-cors-headers==3.2.0
numpy>=1.18.1
pandas==1.0.3
tensorflow==2.0.0
tensorflow-hub==0.8.0
tensorflow_text==2.0.1
xlrd==1.2.0
pymilvus==0.2.10
gunicorn==20.0.4
lmdb==0.98
